Animates our modal on exit.
Once the user accepts our privacy policy, we can dismiss the modal.
Instead of a boring dismissal, we can use animate.css to animate the
modal on exit as well.

With a little bit more stimulus, this is easy. Adding
data-action='modal#submit' connects our Confirm button to the
corresponding stimulus controller function.